requirements.txt
flask
netifaces
python-nmap